+++ demos/default/pages/getStaticProps/env.js
@@ -7,9 +7,7 @@ const Env = ({ env }) => (
     <hr />
     <p>env: {env}</p>
 
-    <Link href="/">
-      <a>Go back home</a>
-    </Link>
+    <Link href="/">Go back home</Link>
   </div>
 )