+++ demos/default/pages/static.js
@@ -17,9 +17,7 @@ const Static = (props) => (
 
     <hr />
 
-    <Link href="/">
-      <a>Go back home</a>
-    </Link>
+    <Link href="/">Go back home</Link>
   </div>
 )