+++ packages/runtime/src/templates/edge/runtime.ts
@@ -33,7 +33,7 @@ export interface RequestData {
     name?: string
     params?: { [key: string]: string }
   }
-  url: URL
+  url: string
   body?: ReadableStream<Uint8Array>
 }
 
@@ -82,7 +82,7 @@ const handler = async (req: Request, context: Context) => {
   const request: RequestData = {
     headers: Object.fromEntries(req.headers.entries()),
     geo,
-    url,
+    url: req.url,
     method: req.method,
     ip: context.ip,
     body: req.body ?? undefined,